+++ gatsby-config.js
@@ -123,5 +123,6 @@ module.exports = {
         acl: null,
       },
     },
+    // 'gatsby-plugin-static-site',
   ],
 };