+++ src/components/custom/footer.tsx
@@ -5,7 +5,8 @@ import { Props } from '../types';
 interface FooterProps extends Props {
   tree: StoryblokNodeTree[];
   countryCode: string;
-  userTypeFromSlug: 'institutional' | 'retail' | 'advanced';
+  userTypeFromSlug: 'institutional' | 'retail';
+  isSalzburg: boolean;
 }
 
 const Footer = 'rcm-footer' as React.ElementType;
@@ -40,7 +41,9 @@ function getCurrentTree(tree: Map<string, any>, lang = 'at-de'): unknown[] {
 }
 
 export const RcmFooter = (props: FooterProps): JSX.Element => {
-  const { tree, countryCode, userTypeFromSlug } = props;
+  const {
+    tree, countryCode, userTypeFromSlug, isSalzburg,
+  } = props;
 
   if (!tree) {
     return null;
@@ -64,6 +67,7 @@ export const RcmFooter = (props: FooterProps): JSX.Element => {
       sitemap-links={JSON.stringify(items)}
       user-type-from-slug={userTypeFromSlug}
       country-code={countryCode}
+      is-salzburg={isSalzburg}
     >
       {/*
        * NOTE: Only works if renderTree is defined using the function keyword!